requirements.txt
streamlit==1.32.0
openai==1.13.3
yfinance==0.2.37
spacy==3.7.4
pytextrank==3.2.5
python-dotenv==1.0.1
langgraph==0.0.15
exa-py==1.7.0
textblob==0.17.1
beautifulsoup4==4.12.3
unstructured==0.12.5
requests==2.31.0
plotly==5.19.0
tensorflow==2.15.0
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
ta==0.11.0